refactor(contract): move clean_tee_status into api/tee_measurements.rs
